+++ sdk/resourcemanager/containerservice/armcontainerservice/response_types.go
@@ -372,13 +372,13 @@ type SnapshotsClientUpdateTagsResponse struct {
 	Snapshot
 }
 
-// TrustedAccessRoleBindingsClientCreateOrUpdateResponse contains the response from method TrustedAccessRoleBindingsClient.CreateOrUpdate.
+// TrustedAccessRoleBindingsClientCreateOrUpdateResponse contains the response from method TrustedAccessRoleBindingsClient.BeginCreateOrUpdate.
 type TrustedAccessRoleBindingsClientCreateOrUpdateResponse struct {
 	// Defines binding between a resource and role
 	TrustedAccessRoleBinding
 }
 
-// TrustedAccessRoleBindingsClientDeleteResponse contains the response from method TrustedAccessRoleBindingsClient.Delete.
+// TrustedAccessRoleBindingsClientDeleteResponse contains the response from method TrustedAccessRoleBindingsClient.BeginDelete.
 type TrustedAccessRoleBindingsClientDeleteResponse struct {
 	// placeholder for future response values
 }
@@ -399,4 +399,10 @@ type TrustedAccessRoleBindingsClientListResponse struct {
 type TrustedAccessRolesClientListResponse struct {
 	// List of trusted access roles
 	TrustedAccessRoleListResult
+}
+
+// UsagesClientListResponse contains the response from method UsagesClient.List.
+type UsagesClientListResponse struct {
+	// The List Usages operation response.
+	ListUsagesResult
 }